+++ example/Form.tsx
@@ -122,13 +122,35 @@ const Form = () => {
               type="checkbox"
             />
           </div>
+        </fieldset>
+        <fieldset className="rdmd-demo--fieldset rdmd-demo--options">
+          <legend>Pipeline</legend>
           <div>
-            <label htmlFor="legacy">Use RDMD (legacy)</label>
-            <input checked={searchParams.has('legacy')} id="legacy" onChange={onCheck('legacy')} type="checkbox" />
+            <label htmlFor="pipeline-rmdx">RMDX</label>
+            <input
+              checked={searchParams.has('rmdx')}
+              id="pipeline-rmdx"
+              onChange={onCheck('rmdx')}
+              type="checkbox"
+            />
           </div>
           <div>
-            <label htmlFor="mdxish">Use MDXish</label>
-            <input checked={searchParams.has('mdxish')} id="mdxish" onChange={onCheck('mdxish')} type="checkbox" />
+            <label htmlFor="pipeline-legacy">RDMD (legacy)</label>
+            <input
+              checked={searchParams.has('legacy')}
+              id="pipeline-legacy"
+              onChange={onCheck('legacy')}
+              type="checkbox"
+            />
+          </div>
+          <div>
+            <label htmlFor="pipeline-mdxish">MDXish</label>
+            <input
+              checked={searchParams.has('mdxish')}
+              id="pipeline-mdxish"
+              onChange={onCheck('mdxish')}
+              type="checkbox"
+            />
           </div>
         </fieldset>
         <textarea name="demo-editor" onChange={onChange} value={doc} />